[parser] Complete re-work to how combat is started and ended.
Previously, I relied on messages from the log about when combat
started and stopped. This meant that I only parsed stuff when I was in
combat -- specifically, if I died or started attacking late then it
wouldn't be accurate. This changes the definition of combat to be a
"contiguous stream of actions without a pause of more than 5 seconds
in damaging abilities. Stated another way, if no one takes damage for
5 seconds then combat will end until someone else damages something.

To make this work for parses against a dummy, we also need to ignore
anyone who isn't in our party. So now it parses the relation field of
the entity ID correctly.